requirements.txt
Django==1.8.2
pytz==2015.4
django-admin-tools==0.5.2
django-debug-toolbar==1.3.2
django-extensions==1.5.5
Werkzeug==0.10.4
django-suit==0.2.13
django-taggit==0.15.0